# swagger_server/models/suscripcion.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Suscripcion(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id_subscription: int=None, id_usuario: int=None, id_plan: int=None, fecha_inicio: date=None, fecha_fin: date=None, estado: str=None):  # noqa: E501
        """Suscripcion - a model defined in Swagger

        :param id_subscription: The id_subscription of this Suscripcion.  # noqa: E501
        :type id_subscription: int
        :param id_usuario: The id_usuario of this Suscripcion.  # noqa: E501
        :type id_usuario: int
        :param id_plan: The id_plan of this Suscripcion.  # noqa: E501
        :type id_plan: int
        :param fecha_inicio: The fecha_inicio of this Suscripcion.  # noqa: E501
        :type fecha_inicio: date
        :param fecha_fin: The fecha_fin of this Suscripcion.  # noqa: E501
        :type fecha_fin: date
        :param estado: The estado of this Suscripcion.  # noqa: E501
        :type estado: str
        """
        self.swagger_types = {
            'id_subscription': int,
            'id_usuario': int,
            'id_plan': int,
            'fecha_inicio': date,
            'fecha_fin': date,
            'estado': str
        }

        self.attribute_map = {
            'id_subscription': 'id_subscription',
            'id_usuario': 'id_usuario',
            'id_plan': 'id_plan',
            'fecha_inicio': 'fecha_inicio',
            'fecha_fin': 'fecha_fin',
            'estado': 'estado'
        }
        self._id_subscription = id_subscription
        self._id_usuario = id_usuario
        self._id_plan = id_plan
        self._fecha_inicio = fecha_inicio
        self._fecha_fin = fecha_fin
        self._estado = estado

    @classmethod
    def from_dict(cls, dikt) -> 'Suscripcion':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Suscripcion of this Suscripcion.  # noqa: E501
        :rtype: Suscripcion
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id_subscription(self) -> int:
        """Gets the id_subscription of this Suscripcion.


        :return: The id_subscription of this Suscripcion.
        :rtype: int
        """
        return self._id_subscription

    @id_subscription.setter
    def id_subscription(self, id_subscription: int):
        """Sets the id_subscription of this Suscripcion.


        :param id_subscription: The id_subscription of this Suscripcion.
        :type id_subscription: int
        """

        self._id_subscription = id_subscription

    @property
    def id_usuario(self) -> int:
        """Gets the id_usuario of this Suscripcion.


        :return: The id_usuario of this Suscripcion.
        :rtype: int
        """
        return self._id_usuario

    @id_usuario.setter
    def id_usuario(self, id_usuario: int):
        """Sets the id_usuario of this Suscripcion.


        :param id_usuario: The id_usuario of this Suscripcion.
        :type id_usuario: int
        """

        self._id_usuario = id_usuario

    @property
    def id_plan(self) -> int:
        """Gets the id_plan of this Suscripcion.


        :return: The id_plan of this Suscripcion.
        :rtype: int
        """
        return self._id_plan

    @id_plan.setter
    def id_plan(self, id_plan: int):
        """Sets the id_plan of this Suscripcion.


        :param id_plan: The id_plan of this Suscripcion.
        :type id_plan: int
        """

        self._id_plan = id_plan

    @property
    def fecha_inicio(self) -> date:
        """Gets the fecha_inicio of this Suscripcion.


        :return: The fecha_inicio of this Suscripcion.
        :rtype: date
        """
        return self._fecha_inicio

    @fecha_inicio.setter
    def fecha_inicio(self, fecha_inicio: date):
        """Sets the fecha_inicio of this Suscripcion.


        :param fecha_inicio: The fecha_inicio of this Suscripcion.
        :type fecha_inicio: date
        """

        self._fecha_inicio = fecha_inicio

    @property
    def fecha_fin(self) -> date:
        """Gets the fecha_fin of this Suscripcion.


        :return: The fecha_fin of this Suscripcion.
        :rtype: date
        """
        return self._fecha_fin

    @fecha_fin.setter
    def fecha_fin(self, fecha_fin: date):
        """Sets the fecha_fin of this Suscripcion.


        :param fecha_fin: The fecha_fin of this Suscripcion.
        :type fecha_fin: date
        """

        self._fecha_fin = fecha_fin

    @property
    def estado(self) -> str:
        """Gets the estado of this Suscripcion.

        Estado de la suscripción (e.g., activa, cancelada, expirada).  # noqa: E501

        :return: The estado of this Suscripcion.
        :rtype: str
        """
        return self._estado

    @estado.setter
    def estado(self, estado: str):
        """Sets the estado of this Suscripcion.

        Estado de la suscripción (e.g., activa, cancelada, expirada).  # noqa: E501

        :param estado: The estado of this Suscripcion.
        :type estado: str
        """

        self._estado = estado