# swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: User API
  version: 1.0.0
servers:
- url: /api
...
paths:
...
  /usuarios/{id_usuario}/suscripcion:
    get:
      tags:
      - Suscripciones
      summary: Obtener el nombre del plan de suscripción de un usuario
      description: Devuelve el nombre del plan de suscripción al que está suscrito
        un usuario específico según el ID del usuario.
      operationId: usuarios_id_usuario_suscripcion_get
      parameters:
      - name: id_usuario
        in: path
        description: ID del usuario
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: Nombre del plan de suscripción del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
        "404":
          description: Usuario no encontrado o sin suscripción activa
      x-openapi-router-controller: swagger_server.controllers.suscripciones_controller
    post:
      tags:
      - Suscripciones
      summary: Crear una nueva suscripción para un usuario
      description: Permite crear una suscripción. La suscripción se añadirá con el
        plan y detalles proporcionados.
      operationId: usuarios_id_usuario_suscripcion_post
      parameters:
      - name: id_usuario
        in: path
        description: ID del usuario para el que se va a crear la suscripción.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Suscripcion'
        required: true
      responses:
        "201":
          description: Suscripción creada con éxito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Suscripcion'
        "400":
          description: "Solicitud incorrecta, faltan datos o parámetros."
      x-openapi-router-controller: swagger_server.controllers.suscripciones_controller
  /suscripciones:
    put:
      tags:
      - Suscripciones
      summary: Actualizar los detalles de la suscripción de usuario
      description: "Permite actualizar los detalles de la suscripción de un usuario,\
        \ como las fechas o el estado de la suscripción."
      operationId: suscripciones_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Suscripcion'
        required: true
      responses:
        "200":
          description: Suscripción de usuario actualizada con éxito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Suscripcion'
        "404":
          description: Suscripción de usuario no encontrada.
      x-openapi-router-controller: swagger_server.controllers.suscripciones_controller
    post:
      tags:
      - Suscripciones
      summary: Crear una nueva suscripción de usuario
      description: Crea una nueva relación entre un usuario y un plan de suscripción.
      operationId: suscripciones_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Suscripcion'
        required: true
      responses:
        "201":
          description: Suscripción de usuario creada con éxito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Suscripcion'
        "400":
          description: "Solicitud incorrecta, faltan datos o parámetros."
      x-openapi-router-controller: swagger_server.controllers.suscripciones_controller
  /suscripciones/{id_subscription}:
    get:
      tags:
      - Suscripciones
      summary: Obtener las suscripciones de un usuario
      description: "Devuelve todas las suscripciones (activas, canceladas, o finalizadas)\
        \ asociadas a un usuario específico según el ID del usuario."
      operationId: suscripciones_id_subscription_get
      parameters:
      - name: id_subscription
        in: path
        description: ID del usuario para obtener todas sus suscripciones.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: Lista de suscripciones del usuario.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Suscripcion'
                x-content-type: application/json
        "404":
          description: Usuario no encontrado o sin suscripciones.
      x-openapi-router-controller: swagger_server.controllers.suscripciones_controller
    delete:
      tags:
      - Suscripciones
      summary: Eliminar una suscripción de usuario
      description: Elimina una relación de suscripción existente entre un usuario
        y un plan de suscripción.
      operationId: suscripciones_id_subscription_delete
      parameters:
      - name: id_subscription
        in: path
        description: ID de la suscripción de usuario a eliminar.
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: Suscripción de usuario eliminada con éxito.
        "404":
          description: Suscripción de usuario no encontrada.
      x-openapi-router-controller: swagger_server.controllers.suscripciones_controller
components:
  schemas:
...
    Suscripcion:
      type: object
      properties:
        id_subscription:
          type: integer
        id_usuario:
          type: integer
        id_plan:
          type: integer
        fecha_inicio:
          type: string
          format: date
        fecha_fin:
          type: string
          format: date
        estado:
          type: string
          description: "Estado de la suscripción (e.g., activa, cancelada, expirada)."
      example:
        estado: estado
        fecha_inicio: 2000-01-23
        id_usuario: 6
        id_subscription: 0
        fecha_fin: 2000-01-23
        id_plan: 1
...
    inline_response_200_1:
      type: object
      properties:
        nombre_plan:
          type: string
          description: Nombre del plan de suscripción
      example:
        nombre_plan: nombre_plan